Восстановлен сборка apk

+++ android/build-apk.sh
@@ -1,27 +1,31 @@
 #!/bin/bash
 
-PATH="$ANDROID/ndk/current:$PATH"       # for 'ndk-build'
-PATH="$ANDROID/ndk/current/simpleperf/bin:$PATH"             # for 'android'
-PATH="$ANDROID/platform-tools:$PATH"    # for 'adb'
-PATH="/home/r/Android/android-studio/jre/bin:$PATH" # for 'java'
-export ANDROID_HOME="$ANDROID"          # for gradle
-export ANDROID_NDK_HOME="$ANDROID/ndk/current"      # for gradle
+[ -z $ANDROID ] && { echo "Set ANDROID environment vairable!"; exit; }
+
+PATH="$ANDROID/ndk:$PATH"                # for 'ndk-build'
+PATH="$ANDROID/ndk/simpleperf/bin:$PATH" # for 'android'
+
+# ./sdkmanager --sdk_root=$ANDROID/sdk
+# openjdk version "1.8.0_242" из Ubuntu
+PATH="$ANDROID/sdk/platform-tools:$PATH"    # for 'adb'
+export ANDROID_HOME="$ANDROID/sdk"          # for gradle
+export ANDROID_NDK_HOME="$ANDROID/ndk"      # for gradle
 
 PROJECT_ROOT="$( cd "$( dirname "${BASH_SOURCE[0]}" )/.." >/dev/null 2>&1 && pwd )"
+rm -rf $PROJECT_ROOT/SDL/build/github.rokashevich.cp19
 
 cd $PROJECT_ROOT/SDL/build-scripts
-rm -rf $PROJECT_ROOT/SDL/build/github.rokashevich.cp19
-./androidbuild.sh github.rokashevich.cp19 $PROJECT_ROOT/src/main.cpp
+read -p "android build"
+COPYSOURCE=0 ./androidbuild.sh github.rokashevich.cp19 $PROJECT_ROOT/src/main.cpp $PROJECT_ROOT/src/gameworld.cpp
 
 cd $PROJECT_ROOT/SDL/build/github.rokashevich.cp19/app/jni/src
-rm *.cpp *.h
-ln -s $PROJECT_ROOT/src/*.cpp .
-ln -s $PROJECT_ROOT/src/*.h .
-
+rm *.cpp
+ln -s $PROJECT_ROOT/src/* .
 cd $PROJECT_ROOT/SDL/build/github.rokashevich.cp19/app/jni
 ln -s $PROJECT_ROOT/glm .
-
+read -p "dbg"
 cp -rv $PROJECT_ROOT/android/SDL/* $PROJECT_ROOT/SDL
 cd $PROJECT_ROOT/SDL/build/github.rokashevich.cp19
+read -p "gradlew"
 ./gradlew assembleDebug
